refactor: fix compile errors
