+++ web-client/src/models/posts/PostFirestoreConverter.ts
@@ -1,10 +1,10 @@
 import { IPost } from './IPost';
 import { Post } from './Post';
 
-
 export const PostFirestoreConverter: firebase.firestore.FirestoreDataConverter<Post> = {
   fromFirestore: (
-    data: firebase.firestore.QueryDocumentSnapshot<IPost>
+    data: firebase.firestore.QueryDocumentSnapshot<IPost>,
   ): Post => Post.factory(data.data()),
-  toFirestore: (modelObject: Post): firebase.firestore.DocumentData => modelObject.toObject(),
+  toFirestore: (modelObject: Post): firebase.firestore.DocumentData =>
+    modelObject.toObject(),
 };
