+++ web-client/src/models/users/IUserGeneral.ts
@@ -1,4 +1,3 @@
-
 export interface IUserGeneral {
   displayName: string;
   displayPicture: string | null;
