+++ web-client/src/modules/create/containers/CreatePostContainer.tsx
@@ -7,8 +7,8 @@ import StepTracker from 'src/components/StepTracker/StepTracker';
 import { createOffer } from 'src/ducks/MyOffers/actions';
 import { createRequest } from 'src/ducks/MyRequests/actions';
 import { ProfileState } from 'src/ducks/profile/types';
-import { IPost } from "src/models/IPost";
-import { PostStatus } from "src/models/PostStatus";
+import { IPost } from 'src/models/posts/IPost';
+import { PostStatus } from 'src/models/posts/PostStatus';
 import { IUser } from 'src/models/users';
 import { IUserAddress } from 'src/models/users/privilegedInformation';
 import NewAddressModal from 'src/modules/create/components/NewAddressModal';
